Refactor sorting logic and sort assignment titles using existing "bookSort" function
