+++ src/app/services/gameboardBuilder.ts
@@ -1,46 +1,11 @@
-import {determineAudienceViews, difficultiesOrdered, SortOrder, tags} from "./";
+import {determineAudienceViews, difficultiesOrdered, SortOrder, sortStringsNumerically, tags} from "./";
 import orderBy from "lodash/orderBy";
 import {AudienceContext, ContentSummaryDTO, Difficulty, GameboardDTO, GameboardItem} from "../../IsaacApiTypes";
 import {ContentSummary, Tag} from "../../IsaacAppTypes";
 
-const bookSort = (a: string, b: string) => {
-    const splitRegex = /(\d+)/;
-    const sectionsA = a.split(splitRegex).filter((x) => x != "." && x != "");
-    const sectionsB = b.split(splitRegex).filter((x) => x != "." && x != "");
-
-    for (let i = 0; i < Math.min(sectionsA.length, sectionsB.length); i++) {
-        const isNumberA = sectionsA[i].search(/\d/) != -1;
-        const isNumberB = sectionsB[i].search(/\d/) != -1;
-
-        if (isNumberA && isNumberB) {
-            const numbersA = sectionsA[i].split(/\./).map((x) => parseInt(x));
-            const numbersB = sectionsB[i].split(/\./).map((x) => parseInt(x));
-
-            for (let j = 0; j < Math.min(numbersA.length, numbersB.length); j++) {
-                const comparison = numbersA[j] - numbersB[j];
-                if (comparison) {
-                    return comparison;
-                }
-            }
-
-            if (numbersA.length != numbersB.length) {
-                return numbersB.length - numbersA.length;
-            }
-        } else if (!isNumberA && !isNumberB) {
-            const comparison = sectionsA[i].localeCompare(sectionsB[i]);
-            if (comparison) {
-                return comparison;
-            }
-        } else {
-            return isNumberA ? 1 : -1;
-        }
-    }
-    return sectionsB.length - sectionsA.length;
-};
-
 export const sortQuestions = (sortState: {[s: string]: string}, creationContext?: AudienceContext) => (questions: ContentSummaryDTO[]) => {
     if (sortState["title"] && sortState["title"] != SortOrder.NONE) {
-        const sortedQuestions = questions.sort((a, b) => bookSort(a.title || "", b.title || ""));
+        const sortedQuestions = questions.sort((a, b) => sortStringsNumerically(a.title || "", b.title || ""));
         return sortState["title"] == SortOrder.ASC ? sortedQuestions : sortedQuestions.reverse();
     }
     if (sortState.difficulty && sortState.difficulty != SortOrder.NONE) {
